Modularize desktop auth orchestration
Separate initial session loading, analytics identity state, and cross-window sign-out coordination from the AuthProvider while preserving auth transitions and CloudSync behavior.
